Remove Utils.kt

+++ app/src/main/java/com/boswelja/contactringtonegenerator/ringtonegen/item/ID.kt
@@ -1,6 +1,6 @@
 package com.boswelja.contactringtonegenerator.ringtonegen.item
 
-enum class ID(val id: Long) {
+enum class ID(val id: Int) {
     CONTACT_NAME(0),
     TEXT_ITEM(1)
 }
